// UI/netang/src/app/features/blog-post/models/blog-post.model.ts
import { Category } from "../../category/models/category.model";

export interface BlogPost {
  id: string; 
  title: string;
  shortDescription: string;
  content: string;
  featuredImageUrl: string;
  urlHandle: string;
  author: string;
  publishedDate: Date;
  isVisible: boolean;
  categories: Category[];
}